+++ src/Zeta.SourceGenerators/ObjectContextSchemaFieldGenerator.cs
@@ -34,6 +34,7 @@ internal static class ObjectContextSchemaFieldGenerator
         GenerateNestedObjectOverloads(methods);
         GenerateCollectionOverloads(methods);
         GenerateGenericCollectionOverloads(methods);
+        GenerateDictionaryOverloads(methods);
 
         var fieldMethods = methods.ToString();
         sb.Append(fieldMethods);
@@ -318,5 +319,65 @@ internal static class ObjectContextSchemaFieldGenerator
                                 }
                             """);
         }
+    }
+
+    private static void GenerateDictionaryOverloads(StringBuilder sb)
+    {
+        sb.AppendLine("""
+
+                          /// <summary>
+                          /// Adds a field validator with fluent context-aware schema builder for IDictionary&lt;TKey, TValue&gt; properties.
+                          /// </summary>
+                          public ObjectContextSchema<T, TContext> Field<TKey, TValue>(
+                              Expression<Func<T, System.Collections.Generic.IDictionary<TKey, TValue>>> propertySelector,
+                              Func<DictionaryContextlessSchema<TKey, TValue>, DictionaryContextSchema<TKey, TValue, TContext>> schema)
+                              where TKey : notnull
+                          {
+                              var propertyName = ObjectContextlessSchema<T>.GetPropertyName(propertySelector);
+                              var getter = ObjectContextlessSchema<T>.CreateGetter(propertySelector);
+                              return AddField(new FieldContextContextValidator<T, System.Collections.Generic.IDictionary<TKey, TValue>, TContext>(propertyName, getter, schema(Z.Dictionary<TKey, TValue>())));
+                          }
+
+                          /// <summary>
+                          /// Adds a field validator with fluent contextless schema builder for IDictionary&lt;TKey, TValue&gt; properties.
+                          /// </summary>
+                          public ObjectContextSchema<T, TContext> Field<TKey, TValue>(
+                              Expression<Func<T, System.Collections.Generic.IDictionary<TKey, TValue>>> propertySelector,
+                              Func<DictionaryContextlessSchema<TKey, TValue>, DictionaryContextlessSchema<TKey, TValue>> schema)
+                              where TKey : notnull
+                          {
+                              var propertyName = ObjectContextlessSchema<T>.GetPropertyName(propertySelector);
+                              var getter = ObjectContextlessSchema<T>.CreateGetter(propertySelector);
+                              var configuredSchema = schema(Z.Dictionary<TKey, TValue>());
+                              return AddField(new FieldContextContextValidator<T, System.Collections.Generic.IDictionary<TKey, TValue>, TContext>(propertyName, getter, configuredSchema.Using<TContext>()));
+                          }
+
+                          /// <summary>
+                          /// Adds a field validator with fluent context-aware schema builder for Dictionary&lt;TKey, TValue&gt; properties.
+                          /// </summary>
+                          public ObjectContextSchema<T, TContext> Field<TKey, TValue>(
+                              Expression<Func<T, System.Collections.Generic.Dictionary<TKey, TValue>>> propertySelector,
+                              Func<DictionaryContextlessSchema<TKey, TValue>, DictionaryContextSchema<TKey, TValue, TContext>> schema)
+                              where TKey : notnull
+                          {
+                              var propertyName = ObjectContextlessSchema<T>.GetPropertyName(propertySelector);
+                              var getter = ObjectContextlessSchema<T>.CreateGetter(propertySelector);
+                              return AddField(new FieldContextContextValidator<T, System.Collections.Generic.IDictionary<TKey, TValue>, TContext>(propertyName, getter, schema(Z.Dictionary<TKey, TValue>())));
+                          }
+
+                          /// <summary>
+                          /// Adds a field validator with fluent contextless schema builder for Dictionary&lt;TKey, TValue&gt; properties.
+                          /// </summary>
+                          public ObjectContextSchema<T, TContext> Field<TKey, TValue>(
+                              Expression<Func<T, System.Collections.Generic.Dictionary<TKey, TValue>>> propertySelector,
+                              Func<DictionaryContextlessSchema<TKey, TValue>, DictionaryContextlessSchema<TKey, TValue>> schema)
+                              where TKey : notnull
+                          {
+                              var propertyName = ObjectContextlessSchema<T>.GetPropertyName(propertySelector);
+                              var getter = ObjectContextlessSchema<T>.CreateGetter(propertySelector);
+                              var configuredSchema = schema(Z.Dictionary<TKey, TValue>());
+                              return AddField(new FieldContextContextValidator<T, System.Collections.Generic.IDictionary<TKey, TValue>, TContext>(propertyName, getter, configuredSchema.Using<TContext>()));
+                          }
+                      """);
     }
 }